ASYLUM: fix more GCC warnings (overflow in implicit constant conversion)
- minor whitespace cleanup as well

git-svn-id: http://asylumengine.googlecode.com/svn/trunk@482 0bfb4aae-4ea4-11de-8d8d-752d95cf3e3c

+++ engines/asylum/resources/actor.cpp
@@ -259,7 +259,7 @@ void Actor::update() {
 
 						Sound *sound  = _vm->sound();
 						if (sound->isCacheOk())
-							sound->playMusic(getScene()->getResourcePack(), kResourceMusic_80020001);
+							sound->playMusic(getScene()->getResourcePack(), (int)kResourceMusic_80020001);
 
 						getWorld()->musicCurrentResourceId = 1;
